requirements.txt
python-docx
ollama
streamlit
google-generativeai